refactor(azure): swap AzureContainerInstance and AzureGroupContainer labels
The two labels were inverted relative to Azure's own concepts: the node modeling
the container *group* (with ip_address, subnet attachment, tags) was labeled
AzureContainerInstance, while the node modeling each individual container was
labeled AzureGroupContainer. This made cross-cloud :Container queries land on
the wrong target without an extra hop.

After the swap:
- AzureContainerInstance models the individual container and carries :Container
  (consistent with ECSContainer, KubernetesContainer).
- AzureGroupContainer models the container group and carries no ontology label
  (consistent with ECSTask, KubernetesPod).

The compatibility migration azure_container_label_swap_migration.json renames
existing nodes via a temporary label to avoid collisions and strips any stale
:Container left on groups.
