+++ cartography/data/jobs/analysis/azure_container_label_swap_migration.json
@@ -1,0 +1,25 @@
+{
+  "statements": [
+    {
+      "__comment__": "Step 1: mark nodes currently labeled AzureContainerInstance (the container group in the old schema) with a temporary label so they can be distinguished from the soon-to-be-renamed individual containers.",
+      "query": "MATCH (n:AzureContainerInstance) REMOVE n:AzureContainerInstance SET n:_AzureGroupSwapTmp",
+      "iterative": false
+    },
+    {
+      "__comment__": "Step 2: rename nodes currently labeled AzureGroupContainer (the individual container in the old schema) to AzureContainerInstance, which now labels the individual container.",
+      "query": "MATCH (n:AzureGroupContainer) REMOVE n:AzureGroupContainer SET n:AzureContainerInstance",
+      "iterative": false
+    },
+    {
+      "__comment__": "Step 3: promote the temporary label to AzureGroupContainer, which now labels the container group.",
+      "query": "MATCH (n:_AzureGroupSwapTmp) REMOVE n:_AzureGroupSwapTmp SET n:AzureGroupContainer",
+      "iterative": false
+    },
+    {
+      "__comment__": "Cleanup: strip stale :Container from AzureGroupContainer. Groups never carry :Container under the new schema; this removes the label from any node that had it before the fad226c8 cleanup ran.",
+      "query": "MATCH (n:AzureGroupContainer:Container) REMOVE n:Container",
+      "iterative": false
+    }
+  ],
+  "name": "Azure container label swap migration"
+}
