+++ cartography/intel/azure/__init__.py
@@ -200,14 +200,14 @@ def _sync_one_subscription(
         update_tag,
         common_job_parameters,
     )
-    container_instances.sync(
+    group_containers.sync_group_containers(
         neo4j_session,
         credentials,
         subscription_id,
         update_tag,
         common_job_parameters,
     )
-    group_containers.sync_group_containers(
+    container_instances.sync(
         neo4j_session,
         credentials,
         subscription_id,
@@ -380,11 +380,12 @@ def start_azure_ingestion(neo4j_session: neo4j.Session, config: Config) -> None:
             common_job_parameters,
         )
 
-        # DEPRECATED: compatibility migration for AzureContainerInstance :Container
-        # labels. The group no longer carries the ontology label (the individual
-        # AzureGroupContainer does). Remove in v1.0.0.
+        # DEPRECATED: compatibility migration that swaps the AzureContainerInstance
+        # and AzureGroupContainer labels so AzureContainerInstance now labels the
+        # individual container and AzureGroupContainer labels the group. Remove in
+        # v1.0.0.
         run_analysis_job(
-            "azure_container_instance_label_migration.json",
+            "azure_container_label_swap_migration.json",
             neo4j_session,
             common_job_parameters,
         )
