+++ cartography/intel/azure/group_containers.py
@@ -8,7 +8,11 @@ from azure.mgmt.containerinstance import ContainerInstanceManagementClient
 
 from cartography.client.core.tx import load
 from cartography.graph.job import GraphJob
+from cartography.intel.azure.util.tag import transform_tags
 from cartography.models.azure.group_container import AzureGroupContainerSchema
+from cartography.models.azure.tags.group_container_tag import (
+    AzureGroupContainerTagsSchema,
+)
 from cartography.util import timeit
 
 from .util.credentials import Credentials
@@ -33,35 +37,42 @@ def get_container_groups(credentials: Credentials, subscription_id: str) -> list
 def transform_group_containers(container_groups: list[dict]) -> list[dict]:
     transformed: list[dict[str, Any]] = []
     for group in container_groups:
-        group_id = group.get("id")
-        for container in group.get("containers", []):
-            image = container.get("image")
-            try:
-                image_digest = image.split("@")[1] if image else None
-            except IndexError:
-                image_digest = None
+        subnet_ids: list[str] = []
+        # Azure SDK as_dict() returns flat structure (not nested under "properties")
+        for subnet_ref in group.get(
+            "subnet_ids", group.get("properties", {}).get("subnet_ids", [])
+        ):
+            subnet_id = subnet_ref.get("id")
+            if subnet_id:
+                subnet_ids.append(subnet_id)
 
-            resources = container.get("resources", {})
-            requests = resources.get("requests", {})
-            limits = resources.get("limits", {})
-
-            transformed.append(
-                {
-                    "id": f"{group_id}/{container.get('name')}",
-                    "name": container.get("name"),
-                    "group_id": group_id,
-                    "image": image,
-                    "image_digest": image_digest,
-                    # ACI does not expose host architecture via its API, and ARM64 support
-                    # is not yet GA. All ACI workloads run on amd64 hosts.
-                    "architecture": "amd64",
-                    "architecture_normalized": "amd64",
-                    "cpu_request": requests.get("cpu"),
-                    "memory_request_gb": requests.get("memory_in_gb"),
-                    "cpu_limit": limits.get("cpu"),
-                    "memory_limit_gb": limits.get("memory_in_gb"),
-                },
-            )
+        transformed.append(
+            {
+                "id": group.get("id"),
+                "name": group.get("name"),
+                "location": group.get("location"),
+                "type": group.get("type"),
+                "provisioning_state": group.get(
+                    "provisioning_state",
+                    group.get("properties", {}).get("provisioning_state"),
+                ),
+                "ip_address": (
+                    group.get("ip_address")
+                    or group.get("properties", {}).get("ip_address")
+                    or {}
+                ).get("ip"),
+                "ip_address_type": (
+                    group.get("ip_address")
+                    or group.get("properties", {}).get("ip_address")
+                    or {}
+                ).get("type"),
+                "os_type": group.get(
+                    "os_type", group.get("properties", {}).get("os_type")
+                ),
+                "tags": group.get("tags"),
+                "SUBNET_IDS": subnet_ids,
+            },
+        )
     return transformed
 
 
@@ -82,12 +93,38 @@ def load_group_containers(
 
 
 @timeit
+def load_group_container_tags(
+    neo4j_session: neo4j.Session,
+    subscription_id: str,
+    groups: list[dict],
+    update_tag: int,
+) -> None:
+    tags = transform_tags(groups, subscription_id)
+    load(
+        neo4j_session,
+        AzureGroupContainerTagsSchema(),
+        tags,
+        lastupdated=update_tag,
+        AZURE_SUBSCRIPTION_ID=subscription_id,
+    )
+
+
+@timeit
 def cleanup_group_containers(
     neo4j_session: neo4j.Session, common_job_parameters: dict
 ) -> None:
     GraphJob.from_node_schema(AzureGroupContainerSchema(), common_job_parameters).run(
         neo4j_session
     )
+
+
+@timeit
+def cleanup_group_container_tags(
+    neo4j_session: neo4j.Session, common_job_parameters: dict
+) -> None:
+    GraphJob.from_node_schema(
+        AzureGroupContainerTagsSchema(), common_job_parameters
+    ).run(neo4j_session)
 
 
 @timeit
@@ -102,4 +139,6 @@ def sync_group_containers(
     raw_groups = get_container_groups(credentials, subscription_id)
     transformed = transform_group_containers(raw_groups)
     load_group_containers(neo4j_session, transformed, subscription_id, update_tag)
+    load_group_container_tags(neo4j_session, subscription_id, transformed, update_tag)
     cleanup_group_containers(neo4j_session, common_job_parameters)
+    cleanup_group_container_tags(neo4j_session, common_job_parameters)
