+++ cartography/models/azure/container_instance.py
@@ -1,9 +1,9 @@
-import logging
 from dataclasses import dataclass
 
 from cartography.models.core.common import PropertyRef
 from cartography.models.core.nodes import CartographyNodeProperties
 from cartography.models.core.nodes import CartographyNodeSchema
+from cartography.models.core.nodes import ExtraNodeLabels
 from cartography.models.core.relationships import CartographyRelProperties
 from cartography.models.core.relationships import CartographyRelSchema
 from cartography.models.core.relationships import LinkDirection
@@ -11,24 +11,23 @@ from cartography.models.core.relationships import make_target_node_matcher
 from cartography.models.core.relationships import OtherRelationships
 from cartography.models.core.relationships import TargetNodeMatcher
 
-logger = logging.getLogger(__name__)
 
-
-# --- Node Definitions ---
 @dataclass(frozen=True)
-class AzureContainerInstanceProperties(CartographyNodeProperties):
+class AzureContainerInstanceNodeProperties(CartographyNodeProperties):
     id: PropertyRef = PropertyRef("id")
     name: PropertyRef = PropertyRef("name")
-    location: PropertyRef = PropertyRef("location")
-    type: PropertyRef = PropertyRef("type")
-    provisioning_state: PropertyRef = PropertyRef("provisioning_state")
-    ip_address: PropertyRef = PropertyRef("ip_address")
-    ip_address_type: PropertyRef = PropertyRef("ip_address_type")
-    os_type: PropertyRef = PropertyRef("os_type")
+    group_id: PropertyRef = PropertyRef("group_id")
+    image: PropertyRef = PropertyRef("image")
+    image_digest: PropertyRef = PropertyRef("image_digest")
+    architecture: PropertyRef = PropertyRef("architecture")
+    architecture_normalized: PropertyRef = PropertyRef("architecture_normalized")
+    cpu_request: PropertyRef = PropertyRef("cpu_request")
+    memory_request_gb: PropertyRef = PropertyRef("memory_request_gb")
+    cpu_limit: PropertyRef = PropertyRef("cpu_limit")
+    memory_limit_gb: PropertyRef = PropertyRef("memory_limit_gb")
     lastupdated: PropertyRef = PropertyRef("lastupdated", set_in_kwargs=True)
 
 
-# --- Relationship Definitions ---
 @dataclass(frozen=True)
 class AzureContainerInstanceToSubscriptionRelProperties(CartographyRelProperties):
     lastupdated: PropertyRef = PropertyRef("lastupdated", set_in_kwargs=True)
@@ -48,33 +47,119 @@ class AzureContainerInstanceToSubscriptionRel(CartographyRelSchema):
 
 
 @dataclass(frozen=True)
-class AzureContainerInstanceToSubnetRelProperties(CartographyRelProperties):
+class AzureGroupContainerToContainerInstanceRelProperties(CartographyRelProperties):
     lastupdated: PropertyRef = PropertyRef("lastupdated", set_in_kwargs=True)
 
 
 @dataclass(frozen=True)
-class AzureContainerInstanceToSubnetRel(CartographyRelSchema):
-    target_node_label: str = "AzureSubnet"
+class AzureGroupContainerToContainerInstanceRel(CartographyRelSchema):
+    target_node_label: str = "AzureGroupContainer"
     target_node_matcher: TargetNodeMatcher = make_target_node_matcher(
-        {"id": PropertyRef("SUBNET_IDS", one_to_many=True)},
+        {"id": PropertyRef("group_id")},
+    )
+    direction: LinkDirection = LinkDirection.INWARD
+    rel_label: str = "CONTAINS"
+    properties: AzureGroupContainerToContainerInstanceRelProperties = (
+        AzureGroupContainerToContainerInstanceRelProperties()
+    )
+
+
+@dataclass(frozen=True)
+class AzureContainerInstanceToECRImageRelProperties(CartographyRelProperties):
+    lastupdated: PropertyRef = PropertyRef("lastupdated", set_in_kwargs=True)
+
+
+@dataclass(frozen=True)
+class AzureContainerInstanceToECRImageRel(CartographyRelSchema):
+    target_node_label: str = "ECRImage"
+    target_node_matcher: TargetNodeMatcher = make_target_node_matcher(
+        {"digest": PropertyRef("image_digest")},
     )
     direction: LinkDirection = LinkDirection.OUTWARD
-    rel_label: str = "ATTACHED_TO"
-    properties: AzureContainerInstanceToSubnetRelProperties = (
-        AzureContainerInstanceToSubnetRelProperties()
+    rel_label: str = "HAS_IMAGE"
+    properties: AzureContainerInstanceToECRImageRelProperties = (
+        AzureContainerInstanceToECRImageRelProperties()
     )
 
 
-# --- Main Schema ---
+@dataclass(frozen=True)
+class AzureContainerInstanceToGitLabContainerImageRelProperties(
+    CartographyRelProperties
+):
+    lastupdated: PropertyRef = PropertyRef("lastupdated", set_in_kwargs=True)
+
+
+@dataclass(frozen=True)
+class AzureContainerInstanceToGitLabContainerImageRel(CartographyRelSchema):
+    target_node_label: str = "GitLabContainerImage"
+    target_node_matcher: TargetNodeMatcher = make_target_node_matcher(
+        {"digest": PropertyRef("image_digest")},
+    )
+    direction: LinkDirection = LinkDirection.OUTWARD
+    rel_label: str = "HAS_IMAGE"
+    properties: AzureContainerInstanceToGitLabContainerImageRelProperties = (
+        AzureContainerInstanceToGitLabContainerImageRelProperties()
+    )
+
+
+@dataclass(frozen=True)
+class AzureContainerInstanceToGCPArtifactRegistryContainerImageRelProperties(
+    CartographyRelProperties
+):
+    lastupdated: PropertyRef = PropertyRef("lastupdated", set_in_kwargs=True)
+
+
+@dataclass(frozen=True)
+class AzureContainerInstanceToGCPArtifactRegistryContainerImageRel(
+    CartographyRelSchema
+):
+    target_node_label: str = "GCPArtifactRegistryContainerImage"
+    target_node_matcher: TargetNodeMatcher = make_target_node_matcher(
+        {"digest": PropertyRef("image_digest")},
+    )
+    direction: LinkDirection = LinkDirection.OUTWARD
+    rel_label: str = "HAS_IMAGE"
+    properties: (
+        AzureContainerInstanceToGCPArtifactRegistryContainerImageRelProperties
+    ) = AzureContainerInstanceToGCPArtifactRegistryContainerImageRelProperties()
+
+
+@dataclass(frozen=True)
+class AzureContainerInstanceToGCPArtifactRegistryPlatformImageRelProperties(
+    CartographyRelProperties
+):
+    lastupdated: PropertyRef = PropertyRef("lastupdated", set_in_kwargs=True)
+
+
+@dataclass(frozen=True)
+class AzureContainerInstanceToGCPArtifactRegistryPlatformImageRel(CartographyRelSchema):
+    target_node_label: str = "GCPArtifactRegistryPlatformImage"
+    target_node_matcher: TargetNodeMatcher = make_target_node_matcher(
+        {"digest": PropertyRef("image_digest")},
+    )
+    direction: LinkDirection = LinkDirection.OUTWARD
+    rel_label: str = "HAS_IMAGE"
+    properties: (
+        AzureContainerInstanceToGCPArtifactRegistryPlatformImageRelProperties
+    ) = AzureContainerInstanceToGCPArtifactRegistryPlatformImageRelProperties()
+
+
 @dataclass(frozen=True)
 class AzureContainerInstanceSchema(CartographyNodeSchema):
     label: str = "AzureContainerInstance"
-    properties: AzureContainerInstanceProperties = AzureContainerInstanceProperties()
+    extra_node_labels: ExtraNodeLabels = ExtraNodeLabels(["Container"])
+    properties: AzureContainerInstanceNodeProperties = (
+        AzureContainerInstanceNodeProperties()
+    )
     sub_resource_relationship: AzureContainerInstanceToSubscriptionRel = (
         AzureContainerInstanceToSubscriptionRel()
     )
     other_relationships: OtherRelationships = OtherRelationships(
         [
-            AzureContainerInstanceToSubnetRel(),
+            AzureGroupContainerToContainerInstanceRel(),
+            AzureContainerInstanceToECRImageRel(),
+            AzureContainerInstanceToGitLabContainerImageRel(),
+            AzureContainerInstanceToGCPArtifactRegistryContainerImageRel(),
+            AzureContainerInstanceToGCPArtifactRegistryPlatformImageRel(),
         ],
     )
