+++ docs/root/modules/azure/schema.md
@@ -1907,39 +1907,39 @@ Representation of an [Azure Kubernetes Service Agent Pool](https://learn.microso
     (AzureKubernetesCluster)-[:HAS_AGENT_POOL]->(:AzureKubernetesAgentPool)
     ```
 
-### AzureContainerInstance
+### AzureGroupContainer
 
-Representation of an [Azure Container Instance](https://learn.microsoft.com/en-us/rest/api/container-instances/container-groups/get). In Azure's API this resource is a *container group* that holds one or more individual containers (modeled as [AzureGroupContainer](#azuregroupcontainer)) — analogous to an ECS Task or Kubernetes Pod rather than an individual container.
+Representation of an [Azure Container Group](https://learn.microsoft.com/en-us/rest/api/container-instances/container-groups/get). In Azure's API this resource is a *container group* that holds one or more individual containers (modeled as [AzureContainerInstance](#azurecontainerinstance)) — analogous to an ECS Task or Kubernetes Pod rather than an individual container.
 
-|**id**| The full resource ID of the Container Instance. |
-|name| The name of the Container Instance. |
-|location| The Azure region where the Container Instance is deployed. |
+|**id**| The full resource ID of the Container Group. |
+|name| The name of the Container Group. |
+|location| The Azure region where the Container Group is deployed. |
 |type| The type of the resource (e.g., `Microsoft.ContainerInstance/containerGroups`). |
-|provisioning_state| The deployment status of the Container Instance (e.g., Succeeded). |
-|ip_address| The public IP address of the Container Instance, if one is assigned. |
-|ip_address_type| The IP type of the Container Instance (`Public` or `Private`) when available. |
-|os_type| The operating system type of the Container Instance (e.g., Linux or Windows). |
+|provisioning_state| The deployment status of the Container Group (e.g., Succeeded). |
+|ip_address| The public IP address of the Container Group, if one is assigned. |
+|ip_address_type| The IP type of the Container Group (`Public` or `Private`) when available. |
+|os_type| The operating system type of the Container Group (e.g., Linux or Windows). |
 
 #### Relationships
 
-- An Azure Container Instance is a resource within an Azure Subscription.
+- An Azure Container Group is a resource within an Azure Subscription.
     ```cypher
-    (AzureSubscription)-[:RESOURCE]->(:AzureContainerInstance)
+    (AzureSubscription)-[:RESOURCE]->(:AzureGroupContainer)
     ```
-- Azure Container Instances can be tagged with Azure Tags.
+- Azure Container Groups can be tagged with Azure Tags.
     ```cypher
-    (AzureContainerInstance)-[:TAGGED]->(AzureTag)
+    (AzureGroupContainer)-[:TAGGED]->(AzureTag)
     ```
-- VNet-integrated Container Instances are attached to a Subnet.
+- VNet-integrated Container Groups are attached to a Subnet.
     ```cypher
-    (AzureContainerInstance)-[:ATTACHED_TO]->(:AzureSubnet)
+    (AzureGroupContainer)-[:ATTACHED_TO]->(:AzureSubnet)
     ```
-- An Azure Container Instance (group) contains one or more AzureGroupContainers.
+- An Azure Container Group contains one or more AzureContainerInstances.
     ```cypher
-    (AzureContainerInstance)-[:CONTAINS]->(:AzureGroupContainer)
+    (AzureGroupContainer)-[:CONTAINS]->(:AzureContainerInstance)
     ```
 
-### AzureGroupContainer
+### AzureContainerInstance
 
 Representation of an individual container within an [Azure Container Group](https://learn.microsoft.com/en-us/rest/api/container-instances/container-groups/get). A container group may run one or more containers — this node models each container separately to enable per-container image tracking.
 
@@ -1964,20 +1964,20 @@ Representation of an individual container within an [Azure Container Group](http
 
 #### Relationships
 
-- An AzureGroupContainer is a resource within an Azure Subscription.
+- An AzureContainerInstance is a resource within an Azure Subscription.
     ```cypher
-    (AzureSubscription)-[:RESOURCE]->(:AzureGroupContainer)
+    (AzureSubscription)-[:RESOURCE]->(:AzureContainerInstance)
     ```
-- An Azure Container Instance (group) contains its AzureGroupContainers.
+- An Azure Container Group contains its AzureContainerInstances.
     ```cypher
-    (AzureContainerInstance)-[:CONTAINS]->(:AzureGroupContainer)
+    (AzureGroupContainer)-[:CONTAINS]->(:AzureContainerInstance)
     ```
-- AzureGroupContainers are linked to the image they run when the image is pinned by digest.
+- AzureContainerInstances are linked to the image they run when the image is pinned by digest.
     ```cypher
-    (AzureGroupContainer)-[:HAS_IMAGE]->(ECRImage)
-    (AzureGroupContainer)-[:HAS_IMAGE]->(GitLabContainerImage)
-    (AzureGroupContainer)-[:HAS_IMAGE]->(GCPArtifactRegistryContainerImage)
-    (AzureGroupContainer)-[:HAS_IMAGE]->(GCPArtifactRegistryPlatformImage)
+    (AzureContainerInstance)-[:HAS_IMAGE]->(ECRImage)
+    (AzureContainerInstance)-[:HAS_IMAGE]->(GitLabContainerImage)
+    (AzureContainerInstance)-[:HAS_IMAGE]->(GCPArtifactRegistryContainerImage)
+    (AzureContainerInstance)-[:HAS_IMAGE]->(GCPArtifactRegistryPlatformImage)
     ```
 
 ### AzureLoadBalancer
